Add ability and requirement to specify custom server URL.

+++ lib/database/database_helper.dart
@@ -8,7 +8,9 @@ abstract class DatabaseHelper {
   }
 
   Future delete() async {
-    return db.delete(table);
+    if (db != null) {
+      return db.delete(table);
+    }
   }
 
   int get schemaVersion;
